+++ netutil.c
@@ -5,7 +5,7 @@
 bool recv_fixed(SOCKET s, char* buf, size_t len) {
     size_t pos = 0;
     do {
-        int recvd = recv(s, buf + pos, len - pos, 0);
+        int recvd = recv(s, buf + pos, (int)(len - pos), 0);
 
         if (recvd <= 0) {
             return false;
@@ -19,7 +19,7 @@ bool recv_fixed(SOCKET s, char* buf, size_t len) {
 bool send_fixed(SOCKET s, const char* buf, size_t len) {
     size_t pos = 0;
     do {
-        int sent = send(s, buf + pos, len - pos, 0);
+        int sent = send(s, buf + pos, (int)(len - pos), 0);
 
         if (sent <= 0) {
             return false;
